requirements.txt
fastapi
uvicorn
facenet-pytorch
pillow
torch
numpy
